add requirements
