+++ control/apps/modu/helper.py
@@ -4,11 +4,9 @@ from __future__ import division
 from os.path import getsize
 from os.path import join
 
-from django.conf import settings
-
 from control.control import settings
 from control.control.base import control_response
-from control.control.base import randomname_maker
+from control.control.base import make_id
 from control.control.err_msg import ModuErrorCode
 from control.control.err_msg import DESCRIBErrorCode
 from control.control.err_msg import DELETEErrorCode
@@ -28,32 +26,7 @@ from .tasks import *
 from control.control.logger import getLogger
 logger = getLogger(__name__)
 
-def make_id(action):
-    while True:
-        distri_id = "%s-%s" % (settings.DISTRI_PREFIX, randomname_maker())
-        partable_id = "%s-%s" % (settings.PARTABLE_PREFIX, randomname_maker())
-        timetable_id = "%s-%s" % (settings.TIMETABLE_PREFIX, randomname_maker())
-        aisdata_id = "%s-%s" % (settings.AISDATA_PREFIX, randomname_maker())
-        signal_id = "%s-%s" % (settings.SIGNAL_PREFIX, randomname_maker())
-        demod_type_id = "%s-%s" % (settings.DEMOD_TYPE_PREFIX, randomname_maker())
-        if action == "distri":
-            if not DistriModel.distri_exist_by_id(distri_id):
-                return distri_id
-        if action == "partable":
-            if not PartableModel.partable_exist_by_id(partable_id):
-                return partable_id
-        if action == "timetable":
-            if not TimetableModel.timetable_exist_by_id(timetable_id):
-                return timetable_id
-        if action == "aisdata":
-            if not AisdataModel.aisdata_exist_by_id(aisdata_id):
-                return aisdata_id
-        if action == "signal":
-            if not SignalModel.signal_exist_by_id(signal_id):
-                return signal_id
-        if action == "demod_type":
-            if not DemodType.demodtype_exist_by_id(demod_type_id):
-                return demod_type_id
+
 
 
 def create_ves_distri(payload):
