+++ requirements.txt
@@ -8,6 +8,7 @@ django-user-accounts == 1.3.1
 djangorestframework == 3.3.3
 gevent == 1.1.1
 greenlet == 0.4.9
-MySQL-python == 1.2.3
+mysql-connector-python == 1.2.4
+MySQL-python == 1.2.4
